Add Context so the handler is passed as props in the signinpage component
Co-authoerd-by: Delmer Reed <delmer814@gmail.com>
