+++ ui/src/utils/auth.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export interface SigninProps {
+  onSignInUser: () => void
+}
+
+export const AuthContext = React.createContext<SigninProps>(null)
